add test table screen
