+++ re_data_ui/src/pages/Tables.tsx
@@ -178,7 +178,11 @@ const Tables: React.FC = (): ReactElement => {
             <div className="bg-white rounded-md px-3 py-4">
               <h3 className="mb-3 text-md font-medium">Tests</h3>
               <div className="grid grid-cols-1 gap-4">
-                <TestsPartial />
+                <TestsPartial
+                  showRunAt
+                  showModel={false}
+                  modelName={model}
+                />
               </div>
             </div>
           </section>
